Style: SharedEmail, Private Comments 폴더 분리
도메인의 사이즈가 증가함에따라 Team에서 폴더를 분리합니다.
