+++ src/main/java/woozlabs/echo/domain/sharedEmail/entity/ShareStatus.java
@@ -1,0 +1,7 @@
+package woozlabs.echo.domain.sharedEmail.entity;
+
+public enum ShareStatus {
+    PRIVATE,
+    TEAM,
+    PUBLIC;
+}
