+++ src/main/java/woozlabs/echo/domain/sharedEmail/repository/PrivateCommentRepository.java
@@ -1,7 +1,7 @@
-package woozlabs.echo.domain.team.repository;
+package woozlabs.echo.domain.sharedEmail.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import woozlabs.echo.domain.team.entity.PrivateComment;
+import woozlabs.echo.domain.sharedEmail.entity.PrivateComment;
 
 import java.util.List;
