+++ src/main/java/woozlabs/echo/domain/sharedEmail/repository/SharedInboxRepository.java
@@ -1,7 +1,7 @@
-package woozlabs.echo.domain.team.repository;
+package woozlabs.echo.domain.sharedEmail.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import woozlabs.echo.domain.team.entity.SharedEmail;
+import woozlabs.echo.domain.sharedEmail.entity.SharedEmail;
 
 import java.util.List;
 import java.util.Optional;
