--- src/main/java/woozlabs/echo/domain/team/entity/ShareStatus.java
@@ -1,7 +1,0 @@
-package woozlabs.echo.domain.team.entity;
-
-public enum ShareStatus {
-    PRIVATE,
-    TEAM,
-    PUBLIC
-}
